Added operate functon
Takes 2 number arguments and an operator.
Calls upon different fucntions for different math operations depending
on the value of the operator

One thing that may need to be fixed in the future, is that the operator
when typed into the console doesn't log as a string and the function needs
the operator to be a string in order to function.